+++ crdt/delta/counter.go
@@ -29,7 +29,7 @@ func NewCounter(replica util.Replica) *Counter {
 	}
 }
 
-func (c *Counter) Value() int {
+func (c *Counter) Value() int64 {
 	return c.pos.Value() - c.neg.Value()
 }
